# Conflicts:
#	features/discovery/home/src/main/java/com/tokopedia/home/beranda/data/mapper/factory/HomeVisitableFactoryImpl.kt
#	features/discovery/home/src/main/java/com/tokopedia/home/beranda/di/module/HomeUseCaseModule.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/model/datamodel/ProductTickerInfoDataModel.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/util/DynamicProductDetailMapper.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/viewholder/ProductTickerInfoViewHolder.kt
#	features/merchant/shop_page/src/test/java/com/tokopedia/shop/pageheader/presentation/ShopPageViewModelTest.kt

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/data/model/response/DetailOccResponse.kt
@@ -38,7 +38,7 @@ data class DetailOccResponse(
 
         @SerializedName("shop_id")
         @Expose
-        val shopId: Int = 0,
+        val shopId: Long = 0,
 
         @SerializedName("shop_type")
         @Expose
@@ -66,11 +66,11 @@ data class DetailOccResponse(
 
         @SerializedName("customer_id")
         @Expose
-        val customerId: Int = 0,
+        val customerId: Long = 0,
 
         @SerializedName("warehouse_id")
         @Expose
-        val warehouseId: Int = 0,
+        val warehouseId: Long = 0,
 
         @SerializedName("is_trade_in")
         @Expose
